+++ src/utils/errors/errorCatalog.error.ts
@@ -35,6 +35,10 @@ export const errorCatalog: Record<string, ErrorDefinition> = {
   },
   [ErrorCode.AccessForbidden]: { status: 403, message: "Access Forbidden" },
   [ErrorCode.InvalidInput]: { status: 405, message: "Invalid Input" },
+  [ErrorCode.TooManyRequests]: {
+    status: 429,
+    message: "Too many requests. Please try again later.",
+  },
   [ErrorCode.SendEmaillError]: { status: 500, message: "Email was not sent" },
   [ErrorCode.NotFound]: {
     status: 404,